Изменил скриптинг. Теперь HTML отдельно, JS - отдельно
