+++ auth.js
@@ -1,33 +1,31 @@
-import { getAuth, GoogleAuthProvider, signInWithPopup, onAuthStateChanged, signOut } from "https://www.gstatic.com/firebasejs/10.12.0/firebase-auth.js";
-import { loadOrCreateProfile, renderGroupList, startChat, setCurrentUser } from './profile.js';
+import { initializeApp } from "https://www.gstatic.com/firebasejs/10.3.0/firebase-app.js";
+import { getAuth, signInWithPopup, GoogleAuthProvider, signOut } from "https://www.gstatic.com/firebasejs/10.3.0/firebase-auth.js";
+import { getFirestore } from "https://www.gstatic.com/firebasejs/10.3.0/firebase-firestore.js";
+import { firebaseConfig } from "./firebase-config.js"
+import { googleLoginBtn, logoutBtn } from "./globals.js"
 
-const auth = getAuth();
-const provider = new GoogleAuthProvider();
+export const app = initializeApp(firebaseConfig);
+export const db = getFirestore(app);
+export const auth = getAuth(app);
 
-export function initAuth() {
-  const loginForm = document.getElementById('loginForm');
-  const mainLayout = document.getElementById('mainLayout');
-  const profileBtn = document.getElementById('profileBtn');
-  const googleLoginBtn = document.getElementById('googleLoginBtn');
-  const logoutBtn = document.getElementById('logoutBtn');
+googleLoginBtn.onclick = () => {
+    loginMsg.textContent = "";
+    signInWithPopup(auth, new GoogleAuthProvider()).catch(
+        (e) => (loginMsg.textContent = "Ошибка входа: " + e.message)
+    );
+};
 
-  onAuthStateChanged(auth, async user => {
-    if (user) {
-      setCurrentUser(user);               // обновляем currentUser в profile.js
-      await loadOrCreateProfile();       // грузим профиль
-      loginForm.style.display = 'none';
-      mainLayout.style.display = 'flex';
-      profileBtn.style.display = 'block';
-      renderGroupList();
-      startChat();
-    } else {
-      loginForm.style.display = 'block';
-      mainLayout.style.display = 'none';
-      profileBtn.style.display = 'none';
+logoutBtn.onclick = () => {
+    if (unsubscribe) {
+        unsubscribe();
+        unsubscribe = null;
     }
-  });
+    signOut(auth);
 
-  googleLoginBtn.onclick = () => signInWithPopup(auth, provider);
-  logoutBtn.onclick = () => signOut(auth);
-}
+    profileBtn.style.display = "none";
+    profilePanel.style.display = "none";
+    openJoinModalBtn.style.display = "none";
+    profileBtn.setAttribute("aria-expanded", "false");
 
+    updateChatInputVisibility();
+};
