+++ firebase-config.js
@@ -1,13 +1,6 @@
-// firebase.js
 import { initializeApp } from "https://www.gstatic.com/firebasejs/10.3.0/firebase-app.js";
-import {
-  getFirestore
-} from "https://www.gstatic.com/firebasejs/10.3.0/firebase-firestore.js";
-import {
-  getAuth
-} from "https://www.gstatic.com/firebasejs/10.3.0/firebase-auth.js";
 
-const firebaseConfig = {
+export const firebaseConfig = {
   apiKey: "AIzaSyC-en4T_Vozvrz7o5dyuYRpZ_4j_ACX3pA",
   authDomain: "abobaserver-49923.firebaseapp.com",
   projectId: "abobaserver-49923",
@@ -16,8 +9,4 @@ const firebaseConfig = {
   appId: "1:364642279962:web:d383373e63e81353d067a3",
 };
 
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app);
-const auth = getAuth(app);
-
-export { app, db, auth };
+initializeApp(firebaseConfig)
